# open-api/tabliodb-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/me": {
      "get": {
        "operationId": "getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me/profile": {
      "patch": {
        "operationId": "updateCurrentUserProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurrentUserProfileUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me/avatar": {
      "post": {
        "operationId": "uploadCurrentUserAvatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "auth"
        ]
      },
      "delete": {
        "operationId": "deleteCurrentUserAvatar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/sign-up": {
      "post": {
        "operationId": "signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto_Output"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCredentialDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto_Output"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/password-reset/request": {
      "post": {
        "operationId": "requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetRequestResponseDto_Output"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/password-reset/confirm": {
      "post": {
        "operationId": "confirmPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetConfirmResponseDto_Output"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/api-keys": {
      "post": {
        "operationId": "createApiKey",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreateResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tabliodb API",
    "description": "Collaborative database schema designer API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tabliodb_access_token"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "AuthUserDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "name": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "cursorColor": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "avatarUrl",
          "cursorColor"
        ],
        "additionalProperties": false
      },
      "CurrentUserResponseDto_Output": {
        "$ref": "#/components/schemas/AuthUserDto_Output"
      },
      "CurrentUserProfileUpdateDto": {
        "type": "object",
        "properties": {
          "cursorColor": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          }
        }
      },
      "SignUpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "LoginResponseDto_Output": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto_Output"
          }
        },
        "required": [
          "accessToken",
          "user"
        ],
        "additionalProperties": false
      },
      "LoginCredentialDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LogoutResponseDto_Output": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          }
        },
        "required": [
          "successful"
        ],
        "additionalProperties": false
      },
      "PasswordResetRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          }
        },
        "required": [
          "email"
        ]
      },
      "PasswordResetRequestResponseDto_Output": {
        "type": "object",
        "properties": {
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "nullable": true
          },
          "resetToken": {
            "type": "string",
            "nullable": true
          },
          "resetUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "successful": {
            "type": "boolean"
          }
        },
        "required": [
          "expiresAt",
          "resetToken",
          "resetUrl",
          "successful"
        ],
        "additionalProperties": false
      },
      "PasswordResetConfirmDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 8
          },
          "token": {
            "type": "string",
            "minLength": 16
          }
        },
        "required": [
          "password",
          "token"
        ]
      },
      "PasswordResetConfirmResponseDto_Output": {
        "type": "object",
        "properties": {
          "revokedSessions": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "successful": {
            "type": "boolean"
          }
        },
        "required": [
          "revokedSessions",
          "successful"
        ],
        "additionalProperties": false
      },
      "ApiKeyCreateDto": {
        "type": "object",
        "properties": {
          "name": {
            "default": "API Key",
            "type": "string",
            "minLength": 1
          },
          "permissions": {
            "default": [
              "all"
            ],
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "all",
                "organization.read",
                "organization.manage",
                "project.create",
                "project.read",
                "project.update",
                "project.delete",
                "project.member.manage",
                "diagram.create",
                "diagram.read",
                "diagram.update",
                "diagram.delete",
                "diagram.comment",
                "snapshot.create",
                "snapshot.read",
                "api-key.manage"
              ]
            }
          }
        }
      },
      "ApiKeyCreateResponseDto_Output": {
        "type": "object",
        "properties": {
          "secret": {
            "type": "string"
          },
          "apiKey": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "name": {
                "type": "string"
              },
              "permissions": {
                "type": "array",
                "items": {
                  "type": "string",
                  "enum": [
                    "all",
                    "organization.read",
                    "organization.manage",
                    "project.create",
                    "project.read",
                    "project.update",
                    "project.delete",
                    "project.member.manage",
                    "diagram.create",
                    "diagram.read",
                    "diagram.update",
                    "diagram.delete",
                    "diagram.comment",
                    "snapshot.create",
                    "snapshot.read",
                    "api-key.manage"
                  ]
                }
              }
            },
            "required": [
              "id",
              "name",
              "permissions"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "secret",
          "apiKey"
        ],
        "additionalProperties": false
      },
...
    }
  }
}